requirements.txt
Django==5.1.5
django-cors-headers==4.6.0 
djangorestframework==3.15.2 
whitenoise==6.8.2
psycopg2-binary==2.9.10
markdown==3.7
django-filter==24.3